+++ DAL/Migrations/20231224050151_InitialMigration.cs
@@ -52,7 +52,7 @@ namespace DAL.Migrations
                 name: "Category",
                 columns: table => new
                 {
-                    ID = table.Column<int>(type: "int", nullable: false)
+                    ID = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     CategoryName = table.Column<string>(type: "nvarchar(max)", nullable: false)
                 },
@@ -65,7 +65,7 @@ namespace DAL.Migrations
                 name: "Item",
                 columns: table => new
                 {
-                    ID = table.Column<int>(type: "int", nullable: false)
+                    ID = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     ItemName = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     ItemUnit = table.Column<string>(type: "nvarchar(max)", nullable: false),
