+++ DAL/Repository/GenericRepository.cs
@@ -40,7 +40,7 @@ namespace DAL.Repository
             return _context.Set<T>().ToList();
         }
 
-        public T GetById(int id)
+        public T GetById(long id)
         {
             return _context.Set<T>().Find(id);
         }
